Change money fields into decimals, because it's mysql

+++ db/migrate/20190417191258_change_funds_current_type_in_fundraisers.rb
@@ -1,7 +1,7 @@
 class ChangeFundsCurrentTypeInFundraisers < ActiveRecord::Migration[5.2]
   def self.up
-    change_column :fundraisers, :funds_current, :float, precision: 20, scale: 2
-    change_column :payments, :amount, :float, precision: 20, scale: 2
+    change_column :fundraisers, :funds_current, :decimal, precision: 20, scale: 2
+    change_column :payments, :amount, :decimal, precision: 20, scale: 2
   end
 
   def self.down
